agent-service/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
requests==2.32.3
langgraph==0.3.11
langchain-openai==0.3.9
pydantic==2.10.6
python-dotenv==1.0.1
